Remove predictor deploy from run test job yaml
Remove predictor deploy from run test job yaml because it would
fail sometimes duo to the predictor not ready in time to serve
the job. There is a separate yaml to deploy the predictor, and
it is used in make run.perf-test. So it is safe to have two
commands to 1. deploy the predictor 2. run performance test,
instead of packing into one yaml.

Also add --ignore-not-found=true otherwise the make command will
stop with an erro when there is no existing resources.

Signed-off-by: Chin Huang <chhuang@us.ibm.com>